Fix ReportsController spec for unauthenticated users
- Add conditional check for analytics directory cleanup to prevent
errors when directory doesn't exist

- Add HTTP 401 status assertion for unauthenticated CSV endpoint access

+++ spec/controllers/reports_controller_spec.rb
@@ -12,7 +12,7 @@ describe ReportsController, type: :request do
   end
 
   after do
-    FileUtils.remove_dir('public/system/analytics')
+    FileUtils.remove_dir('public/system/analytics') if File.directory?('public/system/analytics')
   end
 
   describe 'authenticated course CSV endpoints' do
@@ -257,6 +257,8 @@ describe ReportsController, type: :request do
       it 'does not set the SidekiqJobContext username' do
         expect(SidekiqJobContext).not_to receive(:username=)
         get '/course_csv', params: { course: course.slug }
+        unauthorized = '401'
+        expect(response).to have_http_status(unauthorized)
       end
     end
   end
